refactor(token-usage): promote usage estimator to core module
Move token accounting out of the context_indicator plugin into a shared
core module code_puppy.token_usage so multiple consumers (context_indicator,
statusline, and the upcoming herdr metadata reporter) can use it without
cross-plugin imports.

- Add code_puppy.token_usage with an explicit public __all__ surface.
- Keep context_indicator/usage.py as a thin re-export shim (identity
  preserved for existing import sites and tests).
- Repoint context_indicator/register_callbacks.py and statusline/payload.py
  at the core module.
- Relocate implementation-focused tests to tests/test_token_usage.py; keep
  the plugin rendering/command tests plus one shim-identity compat test in
  tests/plugins/test_context_indicator_plugin.py.

No behaviour change.
