+++ code_puppy/plugins/context_indicator/register_callbacks.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 from typing import List, Tuple
 
 from code_puppy.callbacks import register_callback
-from code_puppy.plugins.context_indicator.usage import (
+from code_puppy.token_usage import (
     ContextUsage,
     get_current_usage,
 )
